+++ kamehouse-shell/bin/common/functions/vlc/vlc-functions.sh
@@ -79,21 +79,27 @@ checkRunningVlcProcess() {
 
 setVlcParams() {
   log.debug "Setting vlc params"
-  local VLC_START_MINIMIZED=false
-  local MUSIC_PLAYLIST_RX=.*${PLAYLISTS_PATH}/music/.*
   local FILE_EXT=${FILE_TO_PLAY: -3}
+  local MUSIC_PLAYLIST_RX=.*${PLAYLISTS_PATH}/music/.*
+  local IS_MUSIC_PLAYLIST=false
 
   if [ "${FILE_EXT}" == "mp3" ]; then
     log.debug "File to play is an mp3"
-    VLC_START_MINIMIZED=true
+    IS_MUSIC_PLAYLIST=true
   fi
   
   if [[ "${FILE_TO_PLAY}" =~ ${MUSIC_PLAYLIST_RX} ]]; then
     log.debug "File to play is a music playlist"
-    VLC_START_MINIMIZED=true
+    IS_MUSIC_PLAYLIST=true
   fi
 
-  if ${VLC_START_MINIMIZED}; then
-    VLC_PARAMS="--qt-start-minimized --qt-system-tray"
+  if ${IS_MUSIC_PLAYLIST}; then
+    if ${VLC_MUSIC_USE_VISUALIZER}; then
+      log.info "Using visualizer filter. The filter needs to be enabled also on vlc player config"
+      VLC_PARAMS="${VLC_PARAMS} --effect-list=spectrum"
+    else
+      log.info "Starting vlc minimized. The visualization filter needs to be disabled on vlc player config"
+      VLC_PARAMS="${VLC_PARAMS} --qt-start-minimized --qt-system-tray"
+    fi
   fi
 }