+++ changelog.md
@@ -8,6 +8,7 @@
 #### v10.01
 
 - Updated default playlists path to .kamehouse/data/playlists and allowed for music playlists to be loaded too
+- Added option to use visualization filter for audio playlists
 - Moved .kamehouse/ files to .kamehouse/config/
 - Moved scripts data to .kamehouse/data/
 - Added option to disable keep alive scripts in kamehouse.cfg